UI improvement: position of votes
* Showing delegate in Wallet Summary (if wallet has voted and Wallet is not a delegate)

* Removed the transaction incoming and outgoing from Wallet Summary

* Added amount of txs at the txs-overview to 'Sent' and 'Received'

* Made the new adjusted parts mobile friendly
